requirements.txt
blinker==1.8.2
cffi==1.17.1
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==11.0.0
pycparser==2.22
pyheif==0.8.0
Werkzeug==3.0.4
